requirements.txt
Django==4.2.24
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0
psutil==5.9.6